refactor: 아이스크림을 그리는 함수 main에서 분리해 새 파일 생성
